Make movie submissions unique
Make the imdb_id field unique at the database level.
This is set once the flow has reached Movie.save() which means the lag
for contacting the API has already occured. Boring. Since the imdb_id is
part of the IMDb link, check to see if a movie already exists with it
from the form.
Fixes #29